consolidate to ops/ple.py
