--- vllm/models/qwen4_exp/nvidia/ops/ple_gate.py
@@ -1,156 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-"""Fused PLE gating and normalization kernel.
-
-Each program processes one token and one HC stream:
-
-  key_n   = grouped_rmsnorm(key, nk, group=H)
-  query_n = grouped_rmsnorm(hidden,       nq,  group=H)
-  d       = dot(key_n, query_n) / sqrt(H)
-  g       = sigmoid(sign(d) * sqrt(clamp_min(|d|, 1e-6)))
-  gated   = g * value
-  normed  = grouped_rmsnorm(gated, ncw, group=H)
-
-Reductions use fp32. Normalized key/query, the gate, and gated output are
-rounded to the input dtype before their consumers.
-"""
-
-import torch
-
-from vllm.platforms import current_platform
-from vllm.triton_utils import tl, triton
-from vllm.utils.torch_utils import direct_register_custom_op
-
-
-@triton.jit
-def _ple_gate_kernel(
-    key_ptr,  # [T, HC*H] with row stride key_rs
-    value_ptr,  # [T, H] with row stride value_rs
-    hidden_ptr,  # [T, HC*H]
-    nk_ptr,  # [HC*H] norm_key weight
-    nq_ptr,  # [HC*H] norm_query weight
-    ncw_ptr,  # [HC*H] norm_conv weight
-    gated_ptr,  # [T, HC*H] out
-    normed_ptr,  # [T, HC*H] out
-    key_rs,
-    value_rs,
-    eps,
-    H: tl.constexpr,
-    HC: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    launch_pdl: tl.constexpr,
-):
-    t = tl.program_id(0)
-    s = tl.program_id(1)
-    lanes = tl.arange(0, BLOCK_H)
-    mask = lanes < H
-    offs = s * H + lanes
-    dtype: tl.constexpr = key_ptr.dtype.element_ty
-
-    if launch_pdl:
-        tl.extra.cuda.gdc_wait()
-
-    k = tl.load(key_ptr + t * key_rs + offs, mask=mask, other=0.0).to(tl.float32)
-    q = tl.load(hidden_ptr + t * HC * H + offs, mask=mask, other=0.0).to(tl.float32)
-    nk = tl.load(nk_ptr + offs, mask=mask, other=0.0).to(tl.float32)
-    nq = tl.load(nq_ptr + offs, mask=mask, other=0.0).to(tl.float32)
-
-    # Match eager materialization at each intermediate tensor boundary.
-    k_n = (k * tl.rsqrt(tl.sum(k * k) / H + eps) * (1.0 + nk)).to(dtype)
-    q_n = (q * tl.rsqrt(tl.sum(q * q) / H + eps) * (1.0 + nq)).to(dtype)
-
-    products = (k_n.to(tl.float32) * q_n.to(tl.float32)).to(dtype)
-    dot = tl.sum(products.to(tl.float32)).to(dtype).to(tl.float32)
-    d = dot / tl.sqrt(float(H))
-    d = d.to(dtype).to(tl.float32)
-    sign = tl.where(d < 0, -1.0, 0.0)
-    sign = tl.where(d > 0, 1.0, sign)
-    magnitude = tl.sqrt(tl.maximum(tl.abs(d), 1e-6)).to(dtype)
-    g = tl.sigmoid(sign * magnitude)
-    g = g.to(dtype).to(tl.float32)
-
-    v = tl.load(value_ptr + t * value_rs + lanes, mask=mask, other=0.0).to(tl.float32)
-    gated = (g * v).to(dtype)
-    gf = gated.to(tl.float32)
-    ncw = tl.load(ncw_ptr + offs, mask=mask, other=0.0).to(tl.float32)
-    normed = gf * tl.rsqrt(tl.sum(gf * gf) / H + eps) * (1.0 + ncw)
-
-    if launch_pdl:
-        tl.extra.cuda.gdc_launch_dependents()
-    tl.store(gated_ptr + t * HC * H + offs, gated, mask=mask)
-    tl.store(normed_ptr + t * HC * H + offs, normed, mask=mask)
-
-
-def _ple_gate(
-    key: torch.Tensor,
-    value: torch.Tensor,
-    hidden: torch.Tensor,
-    norm_key_w: torch.Tensor,
-    norm_query_w: torch.Tensor,
-    norm_conv_w: torch.Tensor,
-    eps: float,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    num_tokens = hidden.shape[0]
-    h = value.shape[-1]
-    hc = hidden.shape[-1] // h
-    assert key.stride(1) == 1 and value.stride(1) == 1
-    assert hidden.is_contiguous()
-    if key.dtype != value.dtype or key.dtype != hidden.dtype:
-        raise ValueError("key, value, and hidden must have the same dtype")
-    if key.dtype not in (torch.bfloat16, torch.float16):
-        raise ValueError("PLE gate supports BF16 and FP16 inputs")
-    gated = torch.empty_like(hidden)
-    normed = torch.empty_like(hidden)
-    _ple_gate_kernel[(num_tokens, hc)](
-        key,
-        value,
-        hidden,
-        norm_key_w,
-        norm_query_w,
-        norm_conv_w,
-        gated,
-        normed,
-        key.stride(0),
-        value.stride(0),
-        eps,
-        H=h,
-        HC=hc,
-        BLOCK_H=triton.next_power_of_2(h),
-        num_warps=4,
-        launch_pdl=current_platform.is_arch_support_pdl(),
-    )
-    return gated, normed
-
-
-def _ple_gate_fake(
-    key: torch.Tensor,
-    value: torch.Tensor,
-    hidden: torch.Tensor,
-    norm_key_w: torch.Tensor,
-    norm_query_w: torch.Tensor,
-    norm_conv_w: torch.Tensor,
-    eps: float,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    return torch.empty_like(hidden), torch.empty_like(hidden)
-
-
-direct_register_custom_op(
-    op_name="qwen4_exp_ple_gate",
-    op_func=_ple_gate,
-    mutates_args=[],
-    fake_impl=_ple_gate_fake,
-)
-
-
-def ple_gate(
-    key: torch.Tensor,
-    value: torch.Tensor,
-    hidden: torch.Tensor,
-    norm_key_w: torch.Tensor,
-    norm_query_w: torch.Tensor,
-    norm_conv_w: torch.Tensor,
-    eps: float,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    return torch.ops.vllm.qwen4_exp_ple_gate(
-        key, value, hidden, norm_key_w, norm_query_w, norm_conv_w, eps
-    )
